feat(build): experimenting with export
I tried to publish what I had so far to Netlify. As it turned out, in order to do this, I needed to do `next export`. But I could not as Nex
tJS did not know about what is /post/[pid].ts. Thus, I needed to next.config.js with mapping. In order to map, I needed to know in advance th
e lits of my blog posts (that are hosted on Contentful). So I need to fetch them first (lib/fetchContent.js). Yeah...